Refactor rust library into crates
