+++ complex objects/cargo/list/src/items.rs
@@ -18,7 +18,7 @@ use std::ptr;
 use time::Timespec;
 use time::now;
 
-use utils::{
+use ffi_utils::{
     string_to_c_char,
     c_char_to_string,
 };
