+++ complex objects/cargo/list/src/lib.rs
@@ -8,24 +8,33 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-use libc::size_t;
+extern crate libc;
+extern crate rusqlite;
+extern crate time;
+extern crate uuid;
+extern crate ffi_utils;
+
 use std::os::raw::{
     c_char,
     c_int
 };
 use std::sync::{
     Arc,
-    Mutex
+    Mutex,
 };
 
 use rusqlite::Connection;
 
-use items::Item;
-use utils::{
+pub mod categories;
+pub mod items;
+
+use categories::Category;
+use ffi_utils::{
     c_char_to_string,
     read_connection,
-    string_to_c_char
 };
+use items::Item;
+
 
 #[derive(Debug)]
 pub struct CategoryManager {
@@ -215,19 +224,6 @@ impl CategoryManager {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct Category {
-    pub id: isize,
-    pub name: String,
-    pub items: Vec<Item>
-}
-
-impl Drop for Category {
-    fn drop(&mut self) {
-        println!("{:?} is being deallocated", self);
-    }
-}
-
 #[no_mangle]
 pub unsafe extern "C" fn get_all_categories(manager: *const Arc<CategoryManager>) -> *mut Vec<Category> {
     let manager = &*manager;
@@ -256,75 +252,4 @@ pub unsafe extern "C" fn category_new(manager: *const Arc<CategoryManager>, name
     let name = c_char_to_string(name);
     let category = Box::new(manager.create_category(name).unwrap());
     Box::into_raw(category)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_destroy(category: *mut Category) {
-    let _ = Box::from_raw(category);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_get_id(category: *const Category) -> c_int {
-    let category = &*category;
-    category.id as c_int
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_get_name(category: *const Category) -> *mut c_char {
-    let category = &*category;
-    string_to_c_char(category.name.clone())
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_get_items(category: *const Category) -> *mut Vec<Item> {
-    let category = &*category;
-    let boxed_items = Box::new(category.items.clone());
-    Box::into_raw(boxed_items)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_items_count(category: *const Category) -> c_int {
-    let category = &*category;
-    category.items.len() as c_int
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_item_at(item_list: *const Vec<Item>, index: size_t) -> *const Item {
-    let item_list = &*item_list;
-    let index = index as usize;
-    let item = Box::new(item_list[index].clone());
-    Box::into_raw(item)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_list_destroy(category_list: *mut Vec<Category>) {
-    let _ = Box::from_raw(category_list);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_list_count(category_list: *const Vec<Category>) -> c_int {
-    let category_list = &*category_list;
-    category_list.len() as c_int
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_list_item_at(category_list: *const Vec<Category>, index: size_t) -> *const Category {
-    let category_list = &*category_list;
-    let index = index as usize;
-    let category = Box::new(category_list[index].clone());
-    Box::into_raw(category)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn add_category(category_list: *mut Vec<Category>, category: *const Category) {
-    let category_list = &mut*category_list;
-    let category = &*category;
-    category_list.push((*category).clone())
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn category_add_item(category: *mut Category, item: *const Item) {
-    let category = &mut*category;
-    let item = &*item;
-    category.items.push((*item).clone());
 }
