+++ complex objects/cargo/logins/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "logins"
+version = "0.1.0"
+authors = ["Emily Toop <etoop@mozilla.com>"]
+
+[target.'cfg(target_os="android")'.dependencies]
+jni = { version = "0.5", default-features = false }
+
+[dependencies]
+time = "0.1.38"
+uuid = { version = "0.4", features = ["v4"] }
+
+[dependencies.ffi-utils]
+path = "../ffi-utils"
+
+[dependencies.rusqlite]
+version = "0.12"
+# System sqlite might be very old.
+features = ["bundled", "limits"]
