+++ complex objects/cargo/src/lib.rs
@@ -8,13 +8,7 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-extern crate libc;
-extern crate rusqlite;
-extern crate time;
-extern crate uuid;
-
-pub mod logins;
-pub mod categories;
-pub mod items;
-pub mod utils;
-pub mod db;
+extern crate ffi_utils;
+extern crate store;
+extern crate logins;
+extern crate list;
