+++ complex objects/cargo/store/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "store"
+version = "0.1.0"
+authors = ["Emily Toop <etoop@mozilla.com>"]
+
+[target.'cfg(target_os="android")'.dependencies]
+jni = { version = "0.5", default-features = false }
+
+[dependencies.rusqlite]
+version = "0.12"
+# System sqlite might be very old.
+features = ["bundled", "limits"]
+
+[dependencies.ffi-utils]
+path = "../ffi-utils"
+
+[dependencies.list]
+path = "../list"
+
+[dependencies.logins]
+path = "../logins"
